usbradioplus_tune::policy
Trait RuntimeEnvironment

pub trait RuntimeEnvironment {
    // Required methods
    fn effective_user_id(&self) -> u32;
    fn command_available(&self, command: &str) -> bool;
    fn stdin_is_terminal(&self) -> bool;
    fn stdout_is_terminal(&self) -> bool;
    fn stderr_is_terminal(&self) -> bool;
    fn terminal_type(&self) -> Option<String>;
}

Operating-system information required by startup policy.

Required Methods

fn effective_user_id(&self) -> u32

Return the effective numeric user identifier.

fn command_available(&self, command: &str) -> bool

Report whether an executable can be invoked through PATH.

fn stdin_is_terminal(&self) -> bool

Report whether standard input is attached to a terminal.

fn stdout_is_terminal(&self) -> bool

Report whether standard output is attached to a terminal.

fn stderr_is_terminal(&self) -> bool

Report whether standard error is attached to a terminal.

fn terminal_type(&self) -> Option<String>

Return the terminal type supplied through TERM.